+++ components/BlogButton/BlogButton.tsx
@@ -1,6 +1,6 @@
-import { UilPen } from '@iconscout/react-unicons';
 import cn from 'classnames';
 import Link from 'next/link';
+import { UilPen } from '../Icons/UilIcons';
 import styles from './blogbutton.module.css';
 
 export default function BlogButton() {